LPS-77967 Move to internal
