+++ modules/apps/wsrp/wsrp-service/src/main/java/com/liferay/wsrp/internal/util/ElementExtensionHelper.java
@@ -12,12 +12,8 @@
  * details.
  */
 
-package com.liferay.wsrp.util;
+package com.liferay.wsrp.internal.util;
 
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
-
-import java.util.Iterator;
 import java.util.List;
 
 import org.apache.axis.message.MessageElement;
@@ -25,22 +21,14 @@ import org.apache.axis.message.MessageElement;
 /**
  * @author Michael C. Han
  */
-public class AttributeExtensionHelper extends BaseExtensionHelper {
+public class ElementExtensionHelper extends BaseExtensionHelper {
 
 	@Override
 	public void addMessageElement(
-		List<MessageElement> messageElements, String name, String value) {
+		List<MessageElement> messageElements, String localPart, String value) {
 
 		MessageElement messageElement = new MessageElement(
-			"http://www.liferay.com/wsrp", "extension");
-
-		try {
-			messageElement.addAttribute(
-				"http://www.liferay.com/wsrp", "name", name);
-		}
-		catch (Exception e) {
-			_log.error(e, e);
-		}
+			"http://www.liferay.com/wsrp", localPart);
 
 		messageElement.setValue(value);
 
@@ -49,16 +37,7 @@ public class AttributeExtensionHelper extends BaseExtensionHelper {
 
 	@Override
 	public String getNameAttribute(MessageElement messageElement) {
-		Iterator<String> iterator = messageElement.getNamespacePrefixes();
-
-		String namespacePrefix = iterator.next();
-
-		String namespaceURI = messageElement.getNamespaceURI(namespacePrefix);
-
-		return messageElement.getAttributeNS(namespaceURI, "name");
+		return messageElement.getName();
 	}
-
-	private static final Log _log = LogFactoryUtil.getLog(
-		AttributeExtensionHelper.class);
 
 }
