+++ modules/apps/wsrp/wsrp-service/src/main/java/com/liferay/wsrp/internal/util/WSRPURLUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.wsrp.util;
+package com.liferay.wsrp.internal.util;
 
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.model.Company;
